Support multiple vk

+++ cmd/virtual-kubelet/internal/commands/root/root.go
@@ -110,7 +110,7 @@ func runRootCommand(ctx context.Context, s *provider.Store, c Opts) error {
 			return errors.Wrap(err, "could not create resource manager")
 		}
 		c.ListenPort = int32(11250 + i)
-		c.MetricsAddr = ":" + strconv.Itoa(int32(10225+i))
+		c.MetricsAddr = ":" + strconv.Itoa(10225+i)
 		apiConfig, err := getAPIConfig(c)
 		if err != nil {
 			return err
